header new js function added
